refactor(orchestration): prompt 分层注入 —— 878 行巨石拆为 8 个可组合模块
D-13 · P0：按 OpenHands/LangGraph 验证过的稳定性分层模式重构系统 prompt。

**改前**：878 行平铺 INSTRUCTIONS 字符串，全部塞在 orchestrator.ts 里。
**改后**：8 个指令模块按稳定性排序注入，STABLE 前缀最大化 cache 命中。

模块结构：
- instructions/language.ts     STABLE  · 输出语言规则（最高优先级）
- instructions/tool-catalog.ts STABLE  · 工具目录 + 描述
- instructions/pipeline.ts    STABLE  · 研究决策链路 + 质量门 + 迭代纪律
- instructions/strategy.ts    STABLE  · 下单流 + 策略协议 + 审批门
- instructions/market.ts      MARKET  · venue 路由 + 多空 + 时效性 + 归因
- instructions/style.ts       STABLE  · 页面上下文 + 语言风格 + 术语翻译
- instructions/divination.ts  COND    · 狐神签规则
- instructions/index.ts               · buildInstructions() 组合引擎

prompt 功能等价（全部 16 个 section 校验通过），typecheck 零错误。
orchestrator.ts: 878 → 93 行。
